Added fix to the Val() function
It was pointed out by René Kraneis that the Val() function was not
properly handling leading white space, or leading positive/negative
signs in numbers. René also provided a fix for this isssue (Thanks!),
and while testing I also realized that the function was not properly
handling fractional numbers. So the string: '12.34(gallons)' would have
returned '12' in stead of '12.34'. This is also fixed with a minor tweak
to René's regex.