version 1.2.5 fixes to date issues.
Problem: When executing the CreateDate function at the end of a month
and creating a date for a month which is shorter than the current month
there was a 'rollover' effect in which the resulting date object would
increase the intended month by one.

Affected functions: CreateDate, CreateDateTime, CreateTime

Thanks again to René Kraneis for not only finding this bug, but
suggesting the fix as well! :)