version 1.2.9 - improved the Round function
Round used to use the toFixed() JS method, which was problematic. Now,
round uses a much more mathematically sound method for rounding to the
nearest N places.

This does however deviate from the way that ColdFusion's Round function
works. So, if you opt not to pass in the second argument (number of
places to round), it will behave just like CF's Round function.

Thanks again to James Moberg for pointing out the problems across
different browsers with the toFixed() JS method.