1.2.11
Tweaked the Round function one more time to be a bit better. It does
essentially the same thing, but instead of the second parameter being
the number of Nths (nearest 10th, nearest 100th, etc.) it now accepts
the number of "decimal places":

1 = nearest 10th
2 = nearest 100th
etc.

I think this is far more intuitive than how it worked before.